refactor : #80 isLogined 함수 utils 폴더로 분리
