+++ project_frontend/src/components/dropdown.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { onLogout } from '../apis';
 import { useNavigate } from 'react-router-dom';
-import { movePath } from '../hooks/movePath';
+import { movePath } from '../utils/movePath';
 
 const DropDownWrapper = styled.div`
 	z-index: 10000;
