+++ project_frontend/src/utils/index.js
@@ -1,0 +1,4 @@
+import { movePath } from './movePath';
+import { isLogined } from './isLogined';
+
+export { movePath, isLogined };
