+++ project_frontend/src/utils/movePath.js
@@ -1,5 +1,3 @@
-import { useNavigate } from 'react-router-dom';
-
 const movePath = (nav, path, query) => {
 	if (query === undefined) nav(path);
 	else nav(path + query);
